defconfigs: xiaomi: Disable Schedhorizon CPU governor
Almost nobody uses this anymore. Our scheduler is good enough and efficient
now. It also makes us of the perf critical API that we've just dropped recently.
So, Schedhorizon won't compile anymore and will defeat the whole purpose of if
that API usage is removed. So, disable the governor all together and put an end
to it.

Signed-off-by: Tashfin Shakeer Rhythm <tashfinshakeerrhythm@gmail.com>