Revert "BACKPORT: cpufreq: schedutil: Avoid using invalid next_freq"
"next_freq" is never set to UINT_MAX. While this change was made in
upstream with commit ecd288429126 ("cpufreq: schedutil: Don't set
next_freq to UINT_MAX"), it also discarded commit 97739501f207
("cpufreq: schedutil: Avoid using invalid next_freq").

So, the current patch was mistakenly added to the tree that just
breaks the intended functionality and is simply proving to be
useless. Revert it. Do it for schedhorizon too while at it.

This reverts commit e982df9aa3dd5 and 0b8b145ee83ab.

Signed-off-by: Tashfin Shakeer Rhythm <tashfinshakeerrhythm@gmail.com>